delete useless files

+++ lab4_2/lib/display.cpp
@@ -79,7 +79,7 @@ void Display::clear() {
 
 void Display::put_string(int x, int y, std::string str) {
     mvwprintw(win, y, x, str.c_str());  // Print string
-    assert(x + str.length() <= cols, "x + str.length() <= cols");
+    assert(x + (int)str.length() <= cols, "x + str.length() <= cols");
     for (int i = 0; i < (int)str.size(); i++) {
         buffer[y][x + i] = str[i];
     }
